Refactor config utilities
