+++ config/utils/get-project-path.ts
@@ -1,0 +1,5 @@
+import { join } from 'path';
+
+export function getProjectPath(...path: string[]): string {
+  return join(__dirname,'..','..', ...path);
+}
