+++ config/webpack/webpack.main.config.ts
@@ -1,8 +1,9 @@
 import { Configuration } from 'webpack';
 
-import { getWebpackConfig } from './webpack.base.config';
+import { getProjectPath } from '../utils/get-project-path';
+import { getWebpackConfig } from './get-webpack-config';
 
-export const mainProcessConfig = getWebpackConfig('main', ({ getProjectPath }) => {
+export const mainProcessConfig = getWebpackConfig('main', () => {
   const config: Configuration = {
     target: 'electron-main',
     entry: {
